backend/requirements.txt
fastapi[standard]
sqlmodel
passlib[argon2]
python-jose[cryptography]
python-multipart
uvicorn